Add server-lite image
Decouple the server runtime from the full CLI package so the server-lite image
can install only the server, backend-common, common, PostgreSQL support, and
uWSGI runtime requirements. Keep existing import paths as compatibility
re-exports while moving shared server context/configuration code into common.

Also make server shell support optional, pin image-relevant dependencies, build
uWSGI in a builder stage, keep compiler tooling out of the final image, and add
a GHCR release workflow matching the node-lite image workflow.
