+++ vantage6-server/server.sh
