+++ vantage6-server/setup.py
@@ -31,8 +31,8 @@ setup(
     packages=find_namespace_packages(),
     python_requires=">=3.10",
     install_requires=[
-        "azure-identity>=1.22.0",
-        "azure-storage-blob>=12.25.1",
+        "azure-identity==1.25.3",
+        "azure-storage-blob==12.28.0",
         "bcrypt==4.0.1",
         "flask==3.1.1",
         "flask-cors==6.0.0",
@@ -42,7 +42,6 @@ setup(
         "flask-marshmallow==0.15.0",
         "flask-socketio==5.5.1",
         "gevent==25.4.2",
-        "ipython==8.10.0",
         "kombu==5.2.4",
         "marshmallow==3.26.2",
         "pyjwt==2.12.1",
@@ -53,15 +52,21 @@ setup(
         "schema==0.7.5",
         "SQLAlchemy==1.4.46",
         "werkzeug==3.1.5",
-        f"vantage6 == {version_ns['__version__']}",
         f"vantage6-common == {version_ns['__version__']}",
+        f"vantage6-backend-common == {version_ns['__version__']}",
     ],
     extras_require={
         "dev": [
             "coverage==6.4.4",
             "black",
             "pre-commit",
-        ]
+        ],
+        "postgres": [
+            "psycopg2-binary==2.9.10",
+        ],
+        "shell": [
+            "ipython==8.10.0",
+        ],
     },
     package_data={
         "vantage6.server": [
