+++ vantage6-server/vantage6/server/configuration_manager.py
@@ -1,0 +1,7 @@
+# Keep these imports as compatibility re-exports for existing server callers
+# while the implementations live in vantage6.common.
+from vantage6.common.server_context import (
+    LOGGING_VALIDATORS,
+    ServerConfiguration,
+    ServerConfigurationManager,
+)
