+++ vantage6-server/vantage6/server/context.py
@@ -1,0 +1,3 @@
+# Keep these imports as compatibility re-exports for existing server callers
+# while the implementations live in vantage6.common.
+from vantage6.common.server_context import BaseServerContext, ServerContext
