+++ vantage6/vantage6/cli/server/common/__init__.py
@@ -5,7 +5,6 @@ from vantage6.common import error, info
 from vantage6.common.globals import APPNAME
 from vantage6.common.docker.addons import remove_container, get_container
 from vantage6.common.context import AppContext
-
 from vantage6.cli.context.server import ServerContext
